--- apps/docs/src/roadmap/index.md
@@ -1,15 +1,0 @@
----
-# https://vitepress.dev/reference/default-theme-home-page
-layout: home
-
-hero:
-  name: Onyx
-  text: Roadmap
-  tagline: Take a look at our roadmap, statistics and general progress for Onyx.
----
-
-<script lang="ts" setup>
-import OnyxRoadmap from "./OnyxRoadmap.vue"
-</script>
-
-<OnyxRoadmap />
